# Benchmarks for node image access multitenancy