#### What type of PR is this?
/kind feature

#### What this PR does / why we need it:
This PR adds benchmarking for the Ensure Secret Pulled Images feature

#### Which issue(s) this PR fixes:
Related to https://github.com/kubernetes/enhancements/issues/2535

#### Special notes for your reviewer:
This PR is built on top of https://github.com/kubernetes/kubernetes/pull/131882 to include the in-mem caching in the benchmarks.

Benchmark results:
The following conclusions are based off benchmarking done on a Intel(R) Xeon(R) Platinum 8171M CPU @ 2.60GHz (8 core) 28GiB RAM machine. The raw benchmark results can be downloaded here:
[pullintent_write.txt](https://github.com/user-attachments/files/21362481/pullintent_write.txt)
[mustattemptpull_miss.txt](https://github.com/user-attachments/files/21362482/mustattemptpull_miss.txt)
[mustattemptpull_hit.txt](https://github.com/user-attachments/files/21362483/mustattemptpull_hit.txt)
[ifnotchanged_write.txt](https://github.com/user-attachments/files/21362484/ifnotchanged_write.txt)
[direct_cachemiss.txt](https://github.com/user-attachments/files/21362486/direct_cachemiss.txt)
[direct_cachehit.txt](https://github.com/user-attachments/files/21362488/direct_cachehit.txt)

**1. comparing direct FS/in-memory authoritative cache access - CACHE HITS**
The results show an average 95% boost in performance while using an authoritative, in-memory cache.

It can be observed that concurrency boosts direct FS access significantly as in direct access, access to the files is unbounded by any locks on this level.
On the other hand, there is a slight drop in performance for the in-memory cache already once the number of concurrent accesses reaches the size of the striped set of locks. However, the in-memory cache is still about 10x faster than direct file accesss while the number of concurrent accesses is 16x times the size of the striped set of locks. 

No new allocations occur for the in-memory authoritative cache (since everything already lives in memory).
```
                                                                       │ FileSystemAccessor │     InMemoryCache-Authoritative     │
                                                                       │       sec/op       │   sec/op     vs base                │
PullRecordsAccessorsCacheHit/RecordsStored=50/ConcurrencyPerCPU=0-8           172.695µ ± 2%   1.385µ ± 4%  -99.20% (p=0.000 n=10)
PullRecordsAccessorsCacheHit/RecordsStored=100/ConcurrencyPerCPU=0-8          171.424µ ± 2%   1.399µ ± 4%  -99.18% (p=0.000 n=10)
PullRecordsAccessorsCacheHit/RecordsStored=200/ConcurrencyPerCPU=0-8          173.811µ ± 2%   1.405µ ± 3%  -99.19% (p=0.000 n=10)
PullRecordsAccessorsCacheHit/RecordsStored=500/ConcurrencyPerCPU=0-8          174.698µ ± 2%   1.462µ ± 5%  -99.16% (p=0.000 n=10)
PullRecordsAccessorsCacheHit/RecordsStored=2000/ConcurrencyPerCPU=0-8         175.668µ ± 1%   1.671µ ± 1%  -99.05% (p=0.000 n=10)
PullRecordsAccessorsCacheHit/RecordsStored=50/ConcurrencyPerCPU=1-8            24.930µ ± 3%   2.057µ ± 1%  -91.75% (p=0.000 n=10)
PullRecordsAccessorsCacheHit/RecordsStored=100/ConcurrencyPerCPU=1-8           25.173µ ± 2%   2.092µ ± 1%  -91.69% (p=0.000 n=10)
PullRecordsAccessorsCacheHit/RecordsStored=200/ConcurrencyPerCPU=1-8           24.900µ ± 2%   2.193µ ± 2%  -91.19% (p=0.000 n=10)
PullRecordsAccessorsCacheHit/RecordsStored=500/ConcurrencyPerCPU=1-8           25.150µ ± 1%   2.223µ ± 2%  -91.16% (p=0.000 n=10)
PullRecordsAccessorsCacheHit/RecordsStored=2000/ConcurrencyPerCPU=1-8          25.318µ ± 2%   2.289µ ± 1%  -90.96% (p=0.000 n=10)
PullRecordsAccessorsCacheHit/RecordsStored=50/ConcurrencyPerCPU=2-8            24.697µ ± 1%   2.034µ ± 1%  -91.76% (p=0.000 n=10)
PullRecordsAccessorsCacheHit/RecordsStored=100/ConcurrencyPerCPU=2-8           25.043µ ± 1%   2.070µ ± 2%  -91.74% (p=0.000 n=10)
PullRecordsAccessorsCacheHit/RecordsStored=200/ConcurrencyPerCPU=2-8           24.963µ ± 1%   2.104µ ± 3%  -91.57% (p=0.000 n=10)
PullRecordsAccessorsCacheHit/RecordsStored=500/ConcurrencyPerCPU=2-8           25.134µ ± 1%   2.179µ ± 1%  -91.33% (p=0.000 n=10)
PullRecordsAccessorsCacheHit/RecordsStored=2000/ConcurrencyPerCPU=2-8          25.402µ ± 1%   2.257µ ± 3%  -91.12% (p=0.000 n=10)
PullRecordsAccessorsCacheHit/RecordsStored=50/ConcurrencyPerCPU=8-8            25.087µ ± 1%   2.177µ ± 2%  -91.32% (p=0.000 n=10)
PullRecordsAccessorsCacheHit/RecordsStored=100/ConcurrencyPerCPU=8-8           25.068µ ± 2%   2.201µ ± 2%  -91.22% (p=0.000 n=10)
PullRecordsAccessorsCacheHit/RecordsStored=200/ConcurrencyPerCPU=8-8           24.777µ ± 1%   2.255µ ± 1%  -90.90% (p=0.000 n=10)
PullRecordsAccessorsCacheHit/RecordsStored=500/ConcurrencyPerCPU=8-8           25.184µ ± 1%   2.316µ ± 2%  -90.80% (p=0.000 n=10)
PullRecordsAccessorsCacheHit/RecordsStored=2000/ConcurrencyPerCPU=8-8          25.270µ ± 1%   2.404µ ± 2%  -90.49% (p=0.000 n=10)
PullRecordsAccessorsCacheHit/RecordsStored=50/ConcurrencyPerCPU=16-8           24.701µ ± 2%   2.243µ ± 2%  -90.92% (p=0.000 n=10)
PullRecordsAccessorsCacheHit/RecordsStored=100/ConcurrencyPerCPU=16-8          25.094µ ± 2%   2.286µ ± 2%  -90.89% (p=0.000 n=10)
PullRecordsAccessorsCacheHit/RecordsStored=200/ConcurrencyPerCPU=16-8          24.975µ ± 2%   2.332µ ± 3%  -90.66% (p=0.000 n=10)
PullRecordsAccessorsCacheHit/RecordsStored=500/ConcurrencyPerCPU=16-8          25.443µ ± 1%   2.392µ ± 2%  -90.60% (p=0.000 n=10)
PullRecordsAccessorsCacheHit/RecordsStored=2000/ConcurrencyPerCPU=16-8         25.187µ ± 1%   2.478µ ± 2%  -90.16% (p=0.000 n=10)
geomean                                                                         36.92µ        2.046µ       -94.46%
```

**2. comparing direct FS/in-memory authoritative cache access - CACHE MISSES**

Here we can see the impact of an additional lock we introduced in order to be able to reliably check if the cache miss is authoritative. While running the benchmark in parallel, file access stays unbounded by any locking, but the cache is now locked both at the LRU primitive and in our code checking if the result was authoritative.

In the real world, accessing the files will always be locked on `PullManager` level as we must make sure writing is safe. Therefore these results are not very conclusive but are an interesting illustration of the impact of synchronization locks.

The actual real-world scenario will be presented later on `MustAttemptImagePull()` reads.

```
                                                                       │ FileSystemAccessor │     InMemoryCache-Authoritative     │
                                                                       │       sec/op       │   sec/op     vs base                │
PullRecordsAccessorCacheMiss/RecordsStored=50/ConcurrencyPerCPU=0-8            27.593µ ± 1%   4.323µ ± 3%  -84.34% (p=0.000 n=10)
PullRecordsAccessorCacheMiss/RecordsStored=100/ConcurrencyPerCPU=0-8           27.769µ ± 2%   4.295µ ± 4%  -84.53% (p=0.000 n=10)
PullRecordsAccessorCacheMiss/RecordsStored=200/ConcurrencyPerCPU=0-8           27.581µ ± 3%   4.324µ ± 4%  -84.32% (p=0.000 n=10)
PullRecordsAccessorCacheMiss/RecordsStored=500/ConcurrencyPerCPU=0-8           27.692µ ± 1%   4.233µ ± 3%  -84.72% (p=0.000 n=10)
PullRecordsAccessorCacheMiss/RecordsStored=2000/ConcurrencyPerCPU=0-8          28.403µ ± 1%   4.267µ ± 4%  -84.98% (p=0.000 n=10)
PullRecordsAccessorCacheMiss/RecordsStored=50/ConcurrencyPerCPU=1-8             3.965µ ± 1%   4.790µ ± 2%  +20.81% (p=0.000 n=10)
PullRecordsAccessorCacheMiss/RecordsStored=100/ConcurrencyPerCPU=1-8            3.998µ ± 2%   4.818µ ± 2%  +20.53% (p=0.000 n=10)
PullRecordsAccessorCacheMiss/RecordsStored=200/ConcurrencyPerCPU=1-8            4.023µ ± 1%   4.774µ ± 2%  +18.68% (p=0.000 n=10)
PullRecordsAccessorCacheMiss/RecordsStored=500/ConcurrencyPerCPU=1-8            3.961µ ± 3%   4.729µ ± 2%  +19.39% (p=0.000 n=10)
PullRecordsAccessorCacheMiss/RecordsStored=2000/ConcurrencyPerCPU=1-8           4.011µ ± 2%   4.790µ ± 2%  +19.44% (p=0.000 n=10)
PullRecordsAccessorCacheMiss/RecordsStored=50/ConcurrencyPerCPU=2-8             3.947µ ± 2%   4.933µ ± 2%  +24.97% (p=0.000 n=10)
PullRecordsAccessorCacheMiss/RecordsStored=100/ConcurrencyPerCPU=2-8            3.978µ ± 1%   4.920µ ± 3%  +23.70% (p=0.000 n=10)
PullRecordsAccessorCacheMiss/RecordsStored=200/ConcurrencyPerCPU=2-8            3.990µ ± 2%   4.940µ ± 2%  +23.81% (p=0.000 n=10)
PullRecordsAccessorCacheMiss/RecordsStored=500/ConcurrencyPerCPU=2-8            4.039µ ± 4%   4.876µ ± 1%  +20.74% (p=0.000 n=10)
PullRecordsAccessorCacheMiss/RecordsStored=2000/ConcurrencyPerCPU=2-8           3.975µ ± 4%   4.947µ ± 2%  +24.42% (p=0.000 n=10)
PullRecordsAccessorCacheMiss/RecordsStored=50/ConcurrencyPerCPU=8-8             3.973µ ± 2%   5.094µ ± 1%  +28.20% (p=0.000 n=10)
PullRecordsAccessorCacheMiss/RecordsStored=100/ConcurrencyPerCPU=8-8            3.977µ ± 1%   5.079µ ± 2%  +27.73% (p=0.000 n=10)
PullRecordsAccessorCacheMiss/RecordsStored=200/ConcurrencyPerCPU=8-8            4.011µ ± 1%   5.114µ ± 2%  +27.52% (p=0.000 n=10)
PullRecordsAccessorCacheMiss/RecordsStored=500/ConcurrencyPerCPU=8-8            3.997µ ± 2%   5.068µ ± 2%  +26.78% (p=0.000 n=10)
PullRecordsAccessorCacheMiss/RecordsStored=2000/ConcurrencyPerCPU=8-8           4.021µ ± 2%   5.093µ ± 3%  +26.66% (p=0.000 n=10)
PullRecordsAccessorCacheMiss/RecordsStored=50/ConcurrencyPerCPU=16-8            4.049µ ± 2%   5.211µ ± 1%  +28.70% (p=0.000 n=10)
PullRecordsAccessorCacheMiss/RecordsStored=100/ConcurrencyPerCPU=16-8           3.975µ ± 1%   5.270µ ± 3%  +32.60% (p=0.000 n=10)
PullRecordsAccessorCacheMiss/RecordsStored=200/ConcurrencyPerCPU=16-8           4.007µ ± 2%   5.289µ ± 4%  +32.00% (p=0.000 n=10)
PullRecordsAccessorCacheMiss/RecordsStored=500/ConcurrencyPerCPU=16-8           4.046µ ± 2%   5.300µ ± 6%  +30.97% (p=0.000 n=10)
PullRecordsAccessorCacheMiss/RecordsStored=2000/ConcurrencyPerCPU=16-8          4.013µ ± 2%   5.305µ ± 3%  +32.21% (p=0.000 n=10)
geomean                                                                         5.892µ        4.859µ       -17.53%
```


**3. comparing MustAttemptImagePull() FS/in-memory authoritative cache access - CACHE HITS**

The results are similar to what we were able to observe in 1., with the exception that concurrency increases performance even in case of in-memory access as the sole record accessing is now wrapped with more common operations that actually occur in code that runs safely concurrently, unlocked.

The performance boost of in-memory cache approach is now around 57%.

```
                                                                               │ FileSystemAccessor │     InMemoryCache-Authoritative     │
                                                                               │       sec/op       │   sec/op     vs base                │
PullManagerMustAttemptImagePullFalse/RecordsStored=50/ConcurrencyPerCPU=0-8            335.6µ ±  2%   119.6µ ± 1%  -64.36% (p=0.000 n=10)
PullManagerMustAttemptImagePullFalse/RecordsStored=100/ConcurrencyPerCPU=0-8           334.6µ ±  2%   121.4µ ± 2%  -63.73% (p=0.000 n=10)
PullManagerMustAttemptImagePullFalse/RecordsStored=200/ConcurrencyPerCPU=0-8           336.6µ ±  1%   121.6µ ± 1%  -63.87% (p=0.000 n=10)
PullManagerMustAttemptImagePullFalse/RecordsStored=500/ConcurrencyPerCPU=0-8           336.5µ ±  1%   123.2µ ± 1%  -63.38% (p=0.000 n=10)
PullManagerMustAttemptImagePullFalse/RecordsStored=2000/ConcurrencyPerCPU=0-8          337.9µ ±  2%   122.3µ ± 2%  -63.80% (p=0.000 n=10)
PullManagerMustAttemptImagePullFalse/RecordsStored=50/ConcurrencyPerCPU=1-8            76.63µ ±  5%   28.20µ ± 1%  -63.20% (p=0.000 n=10)
PullManagerMustAttemptImagePullFalse/RecordsStored=100/ConcurrencyPerCPU=1-8           74.34µ ±  3%   28.40µ ± 2%  -61.79% (p=0.000 n=10)
PullManagerMustAttemptImagePullFalse/RecordsStored=200/ConcurrencyPerCPU=1-8           72.64µ ±  2%   28.43µ ± 1%  -60.86% (p=0.000 n=10)
PullManagerMustAttemptImagePullFalse/RecordsStored=500/ConcurrencyPerCPU=1-8           72.74µ ±  2%   28.05µ ± 2%  -61.44% (p=0.000 n=10)
PullManagerMustAttemptImagePullFalse/RecordsStored=2000/ConcurrencyPerCPU=1-8          71.67µ ±  4%   28.11µ ± 2%  -60.78% (p=0.000 n=10)
PullManagerMustAttemptImagePullFalse/RecordsStored=50/ConcurrencyPerCPU=2-8            62.95µ ±  6%   27.67µ ± 1%  -56.04% (p=0.000 n=10)
PullManagerMustAttemptImagePullFalse/RecordsStored=100/ConcurrencyPerCPU=2-8           61.03µ ±  3%   27.99µ ± 2%  -54.14% (p=0.000 n=10)
PullManagerMustAttemptImagePullFalse/RecordsStored=200/ConcurrencyPerCPU=2-8           60.04µ ±  1%   27.80µ ± 1%  -53.69% (p=0.000 n=10)
PullManagerMustAttemptImagePullFalse/RecordsStored=500/ConcurrencyPerCPU=2-8           59.06µ ±  2%   27.80µ ± 2%  -52.93% (p=0.000 n=10)
PullManagerMustAttemptImagePullFalse/RecordsStored=2000/ConcurrencyPerCPU=2-8          59.04µ ±  3%   27.01µ ± 2%  -54.25% (p=0.000 n=10)
PullManagerMustAttemptImagePullFalse/RecordsStored=50/ConcurrencyPerCPU=8-8            60.51µ ±  8%   28.64µ ± 1%  -52.66% (p=0.000 n=10)
PullManagerMustAttemptImagePullFalse/RecordsStored=100/ConcurrencyPerCPU=8-8           59.73µ ±  6%   28.86µ ± 1%  -51.69% (p=0.000 n=10)
PullManagerMustAttemptImagePullFalse/RecordsStored=200/ConcurrencyPerCPU=8-8           57.20µ ±  1%   28.69µ ± 2%  -49.85% (p=0.000 n=10)
PullManagerMustAttemptImagePullFalse/RecordsStored=500/ConcurrencyPerCPU=8-8           57.82µ ±  2%   28.48µ ± 2%  -50.74% (p=0.000 n=10)
PullManagerMustAttemptImagePullFalse/RecordsStored=2000/ConcurrencyPerCPU=8-8          57.36µ ±  2%   27.95µ ± 1%  -51.28% (p=0.000 n=10)
PullManagerMustAttemptImagePullFalse/RecordsStored=50/ConcurrencyPerCPU=16-8           58.31µ ± 10%   28.58µ ± 1%  -50.99% (p=0.000 n=10)
PullManagerMustAttemptImagePullFalse/RecordsStored=100/ConcurrencyPerCPU=16-8          59.22µ ±  6%   28.59µ ± 1%  -51.73% (p=0.000 n=10)
PullManagerMustAttemptImagePullFalse/RecordsStored=200/ConcurrencyPerCPU=16-8          58.98µ ±  2%   28.61µ ± 2%  -51.49% (p=0.000 n=10)
PullManagerMustAttemptImagePullFalse/RecordsStored=500/ConcurrencyPerCPU=16-8          58.10µ ±  1%   28.48µ ± 2%  -50.98% (p=0.000 n=10)
PullManagerMustAttemptImagePullFalse/RecordsStored=2000/ConcurrencyPerCPU=16-8         58.00µ ±  0%   28.25µ ± 1%  -51.30% (p=0.000 n=10)
geomean                                                                                87.46µ         37.80µ       -56.78%
```

**4. comparing MustAttemptImagePull() FS/in-memory authoritative cache access - CACHE MISSES**

Here we enter a real-world scenario to what was observed in 2. We can see that concurrency no longer causes the direct filesystem to be faster, but we remain at about 58% performance boost in the authoritative in-memory cache approach.

One thing to observe is that cache miss cases are now faster than the cache hits in 3. This occurs because the memory access happens early in the `MustAttemptImagePull()` call, and the function short-circuits if no records were found.
```
                                                                              │ FileSystemAccessor │     InMemoryCache-Authoritative     │
                                                                              │       sec/op       │   sec/op     vs base                │
PullManagerMustAttemptImagePullTrue/RecordsStored=50/ConcurrencyPerCPU=0-8             66.26µ ± 3%   14.47µ ± 3%  -78.16% (p=0.000 n=10)
PullManagerMustAttemptImagePullTrue/RecordsStored=100/ConcurrencyPerCPU=0-8            67.13µ ± 2%   14.55µ ± 2%  -78.33% (p=0.000 n=10)
PullManagerMustAttemptImagePullTrue/RecordsStored=200/ConcurrencyPerCPU=0-8            66.67µ ± 1%   14.69µ ± 2%  -77.96% (p=0.000 n=10)
PullManagerMustAttemptImagePullTrue/RecordsStored=500/ConcurrencyPerCPU=0-8            67.00µ ± 2%   14.79µ ± 3%  -77.93% (p=0.000 n=10)
PullManagerMustAttemptImagePullTrue/RecordsStored=2000/ConcurrencyPerCPU=0-8           67.27µ ± 1%   14.56µ ± 2%  -78.36% (p=0.000 n=10)
PullManagerMustAttemptImagePullTrue/RecordsStored=50/ConcurrencyPerCPU=1-8            19.476µ ± 2%   8.260µ ± 3%  -57.59% (p=0.000 n=10)
PullManagerMustAttemptImagePullTrue/RecordsStored=100/ConcurrencyPerCPU=1-8           19.601µ ± 2%   8.189µ ± 1%  -58.22% (p=0.000 n=10)
PullManagerMustAttemptImagePullTrue/RecordsStored=200/ConcurrencyPerCPU=1-8           19.669µ ± 1%   8.132µ ± 4%  -58.66% (p=0.000 n=10)
PullManagerMustAttemptImagePullTrue/RecordsStored=500/ConcurrencyPerCPU=1-8           19.637µ ± 3%   8.175µ ± 2%  -58.37% (p=0.000 n=10)
PullManagerMustAttemptImagePullTrue/RecordsStored=2000/ConcurrencyPerCPU=1-8          19.522µ ± 2%   8.148µ ± 2%  -58.26% (p=0.000 n=10)
PullManagerMustAttemptImagePullTrue/RecordsStored=50/ConcurrencyPerCPU=2-8            14.886µ ± 4%   7.065µ ± 2%  -52.54% (p=0.000 n=10)
PullManagerMustAttemptImagePullTrue/RecordsStored=100/ConcurrencyPerCPU=2-8           15.010µ ± 1%   6.898µ ± 2%  -54.04% (p=0.000 n=10)
PullManagerMustAttemptImagePullTrue/RecordsStored=200/ConcurrencyPerCPU=2-8           14.795µ ± 3%   7.005µ ± 3%  -52.65% (p=0.000 n=10)
PullManagerMustAttemptImagePullTrue/RecordsStored=500/ConcurrencyPerCPU=2-8           14.962µ ± 1%   6.965µ ± 2%  -53.45% (p=0.000 n=10)
PullManagerMustAttemptImagePullTrue/RecordsStored=2000/ConcurrencyPerCPU=2-8          15.153µ ± 2%   6.936µ ± 4%  -54.23% (p=0.000 n=10)
PullManagerMustAttemptImagePullTrue/RecordsStored=50/ConcurrencyPerCPU=8-8            11.886µ ± 5%   6.007µ ± 3%  -49.46% (p=0.000 n=10)
PullManagerMustAttemptImagePullTrue/RecordsStored=100/ConcurrencyPerCPU=8-8           11.655µ ± 6%   6.085µ ± 5%  -47.79% (p=0.000 n=10)
PullManagerMustAttemptImagePullTrue/RecordsStored=200/ConcurrencyPerCPU=8-8           11.759µ ± 4%   6.125µ ± 3%  -47.91% (p=0.000 n=10)
PullManagerMustAttemptImagePullTrue/RecordsStored=500/ConcurrencyPerCPU=8-8           11.834µ ± 3%   6.199µ ± 7%  -47.62% (p=0.000 n=10)
PullManagerMustAttemptImagePullTrue/RecordsStored=2000/ConcurrencyPerCPU=8-8          11.848µ ± 2%   6.086µ ± 3%  -48.64% (p=0.000 n=10)
PullManagerMustAttemptImagePullTrue/RecordsStored=50/ConcurrencyPerCPU=16-8           11.436µ ± 3%   6.046µ ± 3%  -47.13% (p=0.000 n=10)
PullManagerMustAttemptImagePullTrue/RecordsStored=100/ConcurrencyPerCPU=16-8          11.544µ ± 3%   6.004µ ± 2%  -47.99% (p=0.000 n=10)
PullManagerMustAttemptImagePullTrue/RecordsStored=200/ConcurrencyPerCPU=16-8          11.593µ ± 3%   6.047µ ± 4%  -47.84% (p=0.000 n=10)
PullManagerMustAttemptImagePullTrue/RecordsStored=500/ConcurrencyPerCPU=16-8          11.428µ ± 7%   6.156µ ± 3%  -46.13% (p=0.000 n=10)
PullManagerMustAttemptImagePullTrue/RecordsStored=2000/ConcurrencyPerCPU=16-8         11.201µ ± 6%   6.144µ ± 4%  -45.15% (p=0.000 n=10)
geomean                                                                                19.25µ        7.907µ       -58.92%
```

**5. comparing MustAttemptImagePull() FS/in-memory *non*-authoritative cache access - CACHE HITS**

Non-authoritative cache reads are a bit harder to compare because the factor of total records / cache size comes to play. To simplify, for non-authoritative in-memory cache, I chose to stick the number of total records to a constant 2000 and was only changing the ratio of `total records / cache size` to simulate non-authoritative cache being overwhelmed.

From these observations, we can see that a non-authoritative in-memory cache reaches the same performance of direct FS access when the total number of records is at about 4x the size of the in-memory cache, more-or-less regardless of the concurrency level. However, these tests run with randomly generated requests that discard the benefits of the LRU strategy. It is therefore likely that in reality the performance would still be better even at the 4x `total records / cache size` ratio.

```
…