@stlaz 2025-07-21
> > This PR adds benchmarking for the Ensure Secret Pulled Images feature
> 
> Is this PR only about benchmarking? It looks like more than this. Can you please expand the PR description

My bad - this PR is now built on top of https://github.com/kubernetes/kubernetes/pull/131882 to show the full benchmarks. Originally it was vice-versa, I forgot to change the PR description after the change. Fixing it now.