@liggitt 2025-07-22
> /test pull-kubernetes-unit unrelated, looks like unit tests aren't in the best shape at the moment
> 
> ```
> RUN   TestStreamTranslator_WebSocketServerErrors
>     streamtranslator_test.go:766: +# HELP apiserver_stream_translator_requests_total [ALPHA] Total number of requests that were handled by the StreamTranslatorProxy, which processes streaming RemoteCommand/V5
>         +# TYPE apiserver_stream_translator_requests_total counter
>         +apiserver_stream_translator_requests_total{code="400"} 1
>          
> --- FAIL: TestStreamTranslator_WebSocketServerErrors (0.01s)
> ```

Is there an open issue tracking this? cc @seans3 @aojea