@liggitt 2025-07-22
> === CONT  TestParamRef/ParamRefWithLabelSelectorWithNamespaceAllowNotFound
>     admission_test.go:1783: 
>         	Error Trace:	/home/prow/go/src/k8s.io/kubernetes/staging/src/k8s.io/apiserver/pkg/admission/plugin/policy/validating/admission_test.go:1783
>         	            				/home/prow/go/src/k8s.io/kubernetes/staging/src/k8s.io/apiserver/pkg/admission/plugin/policy/validating/admission_test.go:1521
>         	Error:      	Received unexpected error:
>         	            	context deadline exceeded
>         	Test:       	TestParamRef/ParamRefWithLabelSelectorWithNamespaceAllowNotFound
> --- FAIL: TestParamRef/ParamRefWithLabelSelectorWithNamespaceAllowNotFound (4.25s)

Is there an open issue tracking this? @jpbetz @cici37